+++ src/OpenTelemetry.AutoInstrumentation/Instrumentation.cs
@@ -375,6 +375,8 @@ internal static class Instrumentation
                     break;
                 case TracerInstrumentation.Kafka:
                     break;
+                case TracerInstrumentation.RabbitMq:
+                    break;
                 default:
                     Logger.Warning($"Configured trace instrumentation type is not supported: {instrumentation}");
                     if (FailFastSettings.Value.FailFast)